Fixed the function calls in calculate_fbol
I hadn't updated the function calls to the ir and uv correction
functions. I didn't catch this problem because I'm not explicitly
testing the calculate_fbol function. This should change.

Also, the current implementation of the reddening correction routine
needs to change - it neds to convert the magnitudes to fluxes, and then
call the full CCM89 reddening polynomial to convert those fluxes to
their dereddened values. This will allow me to expand the filters I can
use in my code.