Calculates chi-squared for bb_flux fit to data
This required a large refactoring, since the chi-squared statistic needs
to have the uncertainties in the observed fluxes to work. To get those
uncertainties, I needed to go back and change many of the functions so
that they handle not only the observed magnitudes, but they also handle
the uncertainties in those observed magnitudes and correctly propagate
those uncertainties when transforming magnitudes into fluxes, and
calculating things like the quasi-bolometric flux.